fix SongSearch when MusicWheel is filtered
Some MusicWheel SortOrders filter the songs being displayed.  This is at
least, but perhaps not limited to:

 * SortOrder_Preferred
 * SortOrder_Popularity
 * SortOrder_Recent

If one of those^ SortOrders is active on the MusicWheel,
Simply-Love-SM5's SongSearch implementation will correctly set
GAMESTATE's PreferredSong, but after reloading ScreenSelectMusic (the
filtered SortOrder will still be in effect, and) the MusicWheel may not
have the PreferredSong to show.

As a workaround, if one of those SortOrders is active, change the
MusicWheel's SortOrder to Group before reloading SSM.

Addresses issue#659 on Simply-Love-SM5's GitHub.